+++ apps/api/src/domain/games/memory-match/memory-match.repository.interface.ts
@@ -7,7 +7,7 @@ import { MemoryMatchRound } from './models/memory-match-round.entity';
 export const MEMORY_MATCH_REPOSITORY_INJECTION_TOKEN = 'MEMORY_MATCH_REPOSITORY_INJECTION_TOKEN';
 
 export interface IMemoryMatchRepository {
-    create(round: MemoryMatchRound): Promise<InternalError | undefined>;
+    create(round: MemoryMatchRound): Promise<InternalError | AggregateId>;
 
     createMany(rounds: MemoryMatchRound[]): Promise<void>;