+++ apps/api/src/queries/presentation/formatAggregateType.ts
@@ -27,6 +27,12 @@ const resourceTypeAndLabel: AggregateTypeAndLabel = {
     [ResourceType.playlist]: 'Playlist',
 };
 
+/**
+ * Note that we have expanded the type here from `AggregateType` to string. We
+ * are phasing out lookup tables in favor of dynamic registration here. Also, we
+ * want to be robust to unknown `aggregateType`s in general for flexibility
+ * and easy testing (without casting).
+ */
 export default (aggregateType: string): string => {
     const label = resourceTypeAndLabel[aggregateType];